Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -2,8 +2,8 @@
 
 int main()
 {
-    int a;
-    scanf("%d", &a);
+    float a;
+    scanf("%f", &a);
     printf("", a);
 
     return 0;
